// packages/core/src/utils/clients/supabase/types.ts
export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
      EvolutionRun: {
        Row: {
          config: Json
          end_time: string | null
          goal_text: string
          notes: string | null
          run_id: string
          start_time: string
          status: Database["public"]["Enums"]["EvolutionRunStatus"]
        }
        Insert: {
          config: Json
          end_time?: string | null
          goal_text: string
          notes?: string | null
          run_id?: string
          start_time?: string
          status?: Database["public"]["Enums"]["EvolutionRunStatus"]
        }
        Update: {
          config?: Json
          end_time?: string | null
          goal_text?: string
          notes?: string | null
          run_id?: string
          start_time?: string
          status?: Database["public"]["Enums"]["EvolutionRunStatus"]
        }
        Relationships: []
      }
      Generation: {
        Row: {
          best_workflow_version_id: string | null
          comment: string | null
          end_time: string | null
          generation_id: string
          number: number
          run_id: string
          start_time: string
        }
        Insert: {
          best_workflow_version_id?: string | null
          comment?: string | null
          end_time?: string | null
          generation_id?: string
          number: number
          run_id: string
          start_time?: string
        }
        Update: {
          best_workflow_version_id?: string | null
          comment?: string | null
          end_time?: string | null
          generation_id?: string
          number?: number
          run_id?: string
          start_time?: string
        }
        Relationships: [
          {
            foreignKeyName: "fk_generation_best_wfv"
            columns: ["best_workflow_version_id"]
            isOneToOne: false
            referencedRelation: "WorkflowVersion"
            referencedColumns: ["wf_version_id"]
          },
          {
            foreignKeyName: "fk_generation_run"
            columns: ["run_id"]
            isOneToOne: false
            referencedRelation: "EvolutionRun"
            referencedColumns: ["run_id"]
          },
        ]
      }
// ...
      Workflow: {
        Row: {
          created_at: string
          description: string
          updated_at: string
          wf_id: string
        }
        Insert: {
          created_at?: string
          description: string
          updated_at?: string
          wf_id?: string
        }
        Update: {
          created_at?: string
          description?: string
          updated_at?: string
          wf_id?: string
        }
        Relationships: []
      }
// ...
      WorkflowVersion: {
        Row: {
          all_workflow_io: Json[] | null
          commit_message: string
          created_at: string
          dsl: Json
          generation_id: string | null
          iteration_budget: number
          knowledge: Json | null
          operation: Database["public"]["Enums"]["WorkflowOperation"]
          parent_id: string | null
          parent1_id: string | null
          parent2_id: string | null
          time_budget_seconds: number
          updated_at: string
          wf_version_id: string
          workflow_id: string
        }
        Insert: {
          all_workflow_io?: Json[] | null
          commit_message: string
          created_at?: string
          dsl: Json
          generation_id?: string | null
          iteration_budget?: number
          knowledge?: Json | null
          operation?: Database["public"]["Enums"]["WorkflowOperation"]
          parent_id?: string | null
          parent1_id?: string | null
          parent2_id?: string | null
          time_budget_seconds?: number
          updated_at?: string
          wf_version_id?: string
          workflow_id: string
        }
        Update: {
          all_workflow_io?: Json[] | null
          commit_message?: string
          created_at?: string
          dsl?: Json
          generation_id?: string | null
          iteration_budget?: number
          knowledge?: Json | null
          operation?: Database["public"]["Enums"]["WorkflowOperation"]
          parent_id?: string | null
          parent1_id?: string | null
          parent2_id?: string | null
          time_budget_seconds?: number
          updated_at?: string
          wf_version_id?: string
          workflow_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "fk_wfv_generation"
            columns: ["generation_id"]
            isOneToOne: false
            referencedRelation: "Generation"
            referencedColumns: ["generation_id"]
          },
          {
            foreignKeyName: "fk_wfv_parent1"
            columns: ["parent1_id"]
            isOneToOne: false
            referencedRelation: "WorkflowVersion"
            referencedColumns: ["wf_version_id"]
          },
          {
            foreignKeyName: "fk_wfv_parent2"
            columns: ["parent2_id"]
            isOneToOne: false
            referencedRelation: "WorkflowVersion"
            referencedColumns: ["wf_version_id"]
          },
          {
            foreignKeyName: "fk_workflow_version_parent"
            columns: ["parent_id"]
            isOneToOne: false
            referencedRelation: "WorkflowVersion"
            referencedColumns: ["wf_version_id"]
          },
          {
            foreignKeyName: "WorkflowVersion_workflow_id_fkey"
            columns: ["workflow_id"]
            isOneToOne: false
            referencedRelation: "Workflow"
            referencedColumns: ["wf_id"]
          },
        ]
      }
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never

export type TablesUpdate<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Update: infer U
    }
    ? U
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Update: infer U
      }
      ? U
      : never
    : never